better structure for handling lists
## fix handling of checked-ness in items

It turns out the two kinds of lists are more similar than I thought: you
can have a mixed list, and in particular you can have checked items
within an ordered list.

## add structure for list blocks

Create a new kind of block for lists. This isn't working well yet, but
the structure of the code is there now.