reworking the macros to add actual checks
This way, it's not so much of a unit test of the mdast::Node as it is
of the conversion, which is what I really care about.

I still want to make sure I test each Node for the match, but the
results I check should be the MdqNodes that come out.