MdqNode reorg: part 01
- rename `Header` -> `Section`, to better reflect that it's a whole
  section (and not just the `# header text` line)
- make `MdqNode::read` return a `Vec<MdqNode>`, instead of just one.
  This is in anticipation of a larger reorg that formalizes blocks, leaf
  blocks, inlines, etc.

The impetus for this came from trying to improve the Selector logic. I
realized I need a better version of MdqNode: one that is both tighter
constrained for ease of rendering (as I already have with the various
things that take a `Vec<Inline>`, but can also cut across the hierarchy
so that I can do things like selecting individual spans (like list
elements, links, etc).