get rid of `MdqNode::Root` (#41)
Selectors will work on streams of md elements, which they'll flatmap
into other streams of md elements. That means we're going to be working
with iterators of elements (or for now, Vecs of them), meaning that we
don't really need a special element that's just "a stream of elements".

In other words, we don't actually want to think of a doc as a single
element with a list of elements -- we want to think of it as the stream
of those elements.