Selector recursion with MdqNodeRefs
Introduce a new MdqNodeRef, which is basically a souped-up version of
`&MdqNode` that can also peer into sub-portions of the node graph, like
specific ListItems (and soon, things like links).

One advantage this has is that I can easily create slices into the data,
since MdqNodeRefs are cheap, and in particular don't require cloning the
actual MdqNode tree. Without that, slicing into non-MdqNode items
becomes problematic. For example, `MdqNode::List` contains a
`Vec<ListItem>`, where `ListItem` is _not_ an `MdqNode`. Even if I added
a variant `MdqNode::ListItem`, I'd still want `List` to contain the
`ListItem`s specifically, not just a general `Vec<MdqNode>` (I did
experiment with that as well, but it's (a) harder to reason about what
goes where and (b) requires a bunch of `if let`s when I try to render or
select them). That means that if I have a &ListItem ref, I can't just
get its children as a `Vec<MdqNode>`, which means I can't then return
them from a Selector tream. To make that work, I would have to create
some new enum type that's something like
`enum Selectable { Node(&MdqNode), Item(&ListItem)`... which is
basically what this new `MdqNodeRef` is.

To make this work, I had to make the variants checker more complicated,
and specifically move all of it into the macro. That's because I want to
check variants of MdqNodeRef, but that has an explicit lifetime
specifier. Since the checker is static, it would have to be specialized
to `VariantsChecker<MdqNodeRef<'static>>`, which means that
`see(&self, &E)` can't accept an `MdqNodeRef<'a>` (where `'a` is shorter
than `'static`). I couldn't figure out a good way to make this work
without maros, so I just macro'ed it up. It's fine when it works well,
and isn't _too_ bad to debug when it doesn't. Since the bulk of it is
macro'ed up now, I went ahead and made the whole thing macro'ed,
including the `lazy_static!` to declare the variable and the
`#[test] fn all_variants_checked()`.

I don't love this approach, but I've spent too long trying to go back
and forth with others, and I need to make some progress.