add list selector
I've played around with it for manual testing and it seems to work, but
no automated testing yet. Still, this is a big milestone!

For now, there's minimal (basically no) abstraction to unify list
selectors with section selectors. I'll probably do that next, and then
add tests, and then add more selectors.