remove a TODO
Despite the comment, I don't want to remove the `&str` version: it'll be
useful when I want to write a matcher against URLs, for links and images
(and also for alt text for images). But, anywhere that I'm using it to
check for an inline, I should use `matches_inlines` instead.