various renames
- `MdNode` -> `MdElem`, with associated changes to the macros
- `MdqNodeRef` -> `MdElemRef`

This is during #53, but isn't really needed for that ticket. I just
don't really like the "node" aspect, because it doesn't work as nicely
in my head when we're dealing with a stream of them. In other words: yes
it's a tree, but a lot of times the code is thinking less of its tree
structure and more of it as a "thing that can be streamed".