ABANDONED: inline the Paragraph struct
In `tree.md`, turn:

    Paragraph(Paragraph),

into:

    Paragraph(Vec<Inline>),

and rm the corresponding "pub struct Paragraph".

This is ending up being more trouble than it's worth, and I'm not even
sure if it's worth anything. Yeah, I get to remove a struct; but it's a
harmless struct, and it provides a nice visual parallelism with the
other variants that do need a more complex struct.

I'm saving this in case I want to revisit it, since the macros took a
bit of time to get right. But I don't think I'll want it.