add thematic breaks between outputted elements
This makes it a bit visually clearer what's going on.

Also, consolidate thematic breaks: regardless of whether they're
"organic" in the doc, or the ones we're adding as part of this, don't
ever output two in a row.

That may be controversial (maybe someone wants that?) but it'll be good
for now. Eventually I'll probably make this configurable.