add a new MdElemRef::Doc
This lets us enapsulate a `Vec<MdElem>` that represents a single area of
the doc, as opposed to a stream of disparate areas.

This also lets us simplify the SelectResult, since each selector now
returns exactly one element (previously, there was a Multi case for
SectionSelector, but now that's just a single Doc).