flatten the MdElems tree
Inline containers and leaf blocks.

The more precise hierarchy was a nice idea, but it just adds a lot of
extra boilerplate for no real benefit.

I thought this would be helpful for #74, but I don't think it actually
will be. Still, I think it's an improvement, so I'm going to keep it.

A possible next step would be to create a new struct in `tree.rs`:

    pub struct Doc {
        pub body: Vec<MdElem>,
    }

But from trying it, it basically creates a lot more noise in the tests,
and doesn't actually buy much in prod code -- so it's not worth it.