integ tests: move main.rs into lib.rs
Prepare for integ tests (#92) by moving most of `main.rs` into `lib.rs`.
This means integ tests will be able to run things in memory.

Creating the `lib.rs` means that rustdoc actually runs. This broke a
bunch of rustdocs that weren't specifying the right language for code
blocks, so I fixed that.