add simple integ test
This is just an example, but it sets the infrastructure.

I'm writing the test cases in toml for now. Markdown would be pretty
neat, but it presents a bootstrap problem: if my code were broken, the
integ test could break such that it both gets and expects nothing, and
thus passes when it should fail.

That may not be a major concern though, given that I have copious unit
tests. I'm going to experiment with a markdown-based file format, too.

In service of #92.