block-level html elements
HTML can be [block-level or inline][1], but the previous implementation
only allowed inline. It treated block-level HTML as a paragraph with
a single HTML inline element, which has (at least) two problems: it
serializes to JSON incorrectly, and it means that a paragraph search
(#140) would incorrectly include these block-level HTMLs.

This issue will resolve #34, though I still need to add tests after this
commit.

This issue blocks implementation of #153.

[1]: https://github.github.com/gfm/#html-blocks