split up the Selector interface
The only thing the interface actually needs is `fn try_select`. The
`fn matches` is just a convenient way of implementing `try_select`. This
commit formalizes that by making Selector only decaore `fn try_select`,
and without a default impl.

This is in preparation for #141. That implementation will let us create
"table slices", which aren't just a pointer to an existing table, but
actually creates a new table which is a subset of the original. That
won't fit into the `fn matches` pattern, so this PR breaks that pattern
up.