normalize as a separate step
That way, if a table is jagged in the original, we won't normalize it
unless it's used in a selector. In particular, if it's a child of some
other element we're selecting, we'll output it just as it was in the
original.

(I could imagine a flag to always normalize, as we do with links; but
I won't do that in this PR.)