add ability to select tables (#163)
At long last! Table selectors.

Syntax:

```
:-: headers matcher :-: rows matcher
```

- The `headers matcher` _must_ be specified explicitly: if you want "all
columns", use `*`
- Header rows are always matched, regardless of the `rows matcher`
- If the table is jagged (not all rows have the same number of columns),
it will be normalized by extending all short rows with empty columns.
This is a departure from the official spec, which says that the number
of columns is dictated by the header row, and all longer rows are
truncated. I'm making that departure intentionally, because I think it's
nicer, and the output I'll provide will always be valid markdown. I may
later provide a switch to control that behavior.

This resolves #141.