add "chained" integ tests
These are tests of chaining a selector into a selector of the same type,
like `# foo | #foo`. This should always apply a second filter to the
same elements; it basically ANDs them together.

I added these tests for tables initially, since there are separate Table
and TableSlice variants for MdElemRef; this meant that the chaining
needed to be specifically supported, and was no longer a given. But,
then I figured I may as well formalize that.

Some toml test case files don't need this, so there's also a mechanism
to mark the chained tests as unneeded. Otherwise, a test case file
without a case named "chained" will fail (the `build.rs` will
auto-generate a "chained" test that trivially fails).