handle backticks within inline code blocks
An inline code block starts with a string of at least one backtick, but
possibly more: `` ` ``, ` `` `, etc. It then continues until that same
sting again.

The actual surrounding backticks don't matter, as long as they don't
appear in the contents. The following are equivalent, and both produce
`a``b`:

```markdown
`a``b`
```a``b```
```

The easiest way to handle this is to count the longest stretch of backticks in the contents, and then use a string of N+1 backticks for the inline code markers. So, that's what this PR does.

I can imagine nicer formatting, like finding the first available empty spot; or maybe just special-casing ` `` `. But for now, I'm keeping things simple.

If the contents start or end with backspace, we add a single space on either side of the contents (between it and the backticks). The CommonMark spec says these should be stripped on parse; they're currently not, because of newly-filed #171.

This PR resolves #157.