rewrite footnotes (checkpoint)
Mostly working, other than ordering -- the ordering currently comes from
the original text, not the new text.

Rather than having the transformer rewrite directly, I think I should
have it return a string. This makes it easier to work with, and makes
it more efficient. But the lifetimes are tricky.