add the footnote content to context
This wasn't working before because I was trying to be too clever with
the entry. The MdElem::all needs a mutable ref to the ctx, but the entry
was also holding onto one. That was actually correctly identifying a
corner case: that the children could have added a footnote to the
context, thus actually modifying it (maybe even with a reallocation!).
The fix is just to do the check-if-exists separately from the adding.