should WordBoundary::Never actually be OnlyNewlines?
I don't think a newline would actually get written during one of these
spans, but if it does, it looks like a wrap no matter what -- and thus
should probably count as one.

Come to think of it, should newlines be not normal actions, but a custom
action? Or does the caller handle newlines specially if they want to?

In practice, I think we'll want newlines to translate to
Output::ensure_newlines