use `PUSH_LITERAL` to simplify grammar
Use pest's new (as of 2.8.0) `PUSH_LITERAL`. This reduces the number of
rules we need in the grammar.

One complication is that the `string` rule now needs to end with `DROP`.
So we could either keep the `string` as only non-empty and then add the
`DROP` after each `string?`, or we could have it include empty. The
latter is simpler, but means we need to then add a check for explicit
`*` in the table matcher.