read extra positional args as files
Treat the first positional arg as a selector (as before), but any
subsequent ones as files to read. This resolves #275.

From the new docs:

> An optional list of Markdown files to parse, by path. If not provided,
> standard input will be used.
>
> If these are provided, mdq will act as if they were all concatenated
> into a single file. For example, if you
> use --link-pos=doc, the link definitions for all input files will be
> at the very end of the output.
>
> A path of "-" represents standard input.
>
> If these are provided, standard input will not be used unless one of
> the arguments is "-". Files will be
> processed in the order you provide them. If you provide the same file
> twice, mdq will process it twice, unless
> that file is "-"; all but the first "-" paths are ignored.

Now that I have the new `OsFacade`, there's an opportunity to have it
include the output as well, such that we can remove the separate
`run_in_memory` and `run_stdio` implementations. I'll leave that for
#213, though.