split selectors.rs up (#314)
This is just a big mv of code, with no functional change. Other than
wrapping a `Vec<Selector>` into a `SelectorChain` for the sake of a
TryFrom impl, there's literally nothing here other than cut-and-paste,
and corresponding import changes.

This is in service of #302, but does not complete that work. I'm
splitting the file up now, so that it'll be easier to move the files
around later. In particular, I'm going to want to move `selectors.rs`
and the new `matcher.rs` into the `select` package, while keeping the
`*_try_from.rs` files in `query`.