checkpoint: rm SelectorChain
This was an attempt to bring SelectorChain into the Selector mechanism.
Unfortunately, it doesn't quite work. The other Selectors work very
differently; the new code in api.r's find_nodes highlights this (since
the chain is handled as a special case), as does
ChainSelector::try_select, which can't really be implemented.

Part of the problem is that the selector fn for ChainSelector would need
to accept any MdElem. Another problem is that it needs the MdContext,
for find_nodes recursion.