remove SelectorChain struct
Instead of having a separate SelectorChain, create a new variant
Selector::Chain.

The SelectorChain struct had a slightly different contract than
Selectors: it worked on a stream of MdElem (rather than just a 1:1
mapping), and essentially flatmapped it, with the extra caveat that an
empty chain on an MdDoc would not flatmap, but just return the MdDoc
directly. This has now been encoded into ChainSelector::try_select.

resolves #315