expose query string errors a bit
Don't actually expose much about them; basically just a Display.

This is in service of #307. That ticket says to hide pest, but actually
it was already hidden. The closest we got was in exposing
`run::Error::QueryParse { error: ParseError, .. }`, but `ParseError`
itself was not pub.

This makes it a bit hard to work with, so this PR actually exposes it a
bit _more_. The pest aspect remains hidden, but `QueryParse` is now a
pub struct with private fields and a pub `impl Display`.