possible parsing improvements for regexes
These have a good amount of functional testing, so I'm not worried about
that. But I'd be interested to see if there's actually any perf benefit.

The motivation here is that I noticed there are a lot of Pairs being
created (in the code before this PR), since each char in the regex match
is a single-char Pair. This makes debugging a pain, but it also means
there are more heap allocations (I _think_ Pairs are heap-allocated,
but I may be wrong about that).

I should do at least a bit of perf testing before I merge this.