mark various things pub(crate)
Change various items from `pub` to `pub(crate)`. These items weren't
actually public, since their modules weren't exported as pub. This
doesn't change that, but it makes it clearer in the code.

Also add a CI check (and a helper script) to ensure that I'll maintain
this practice going forward.

Resolves #383