handle autolinks in parsing and rendering
The `mdq::md_elem::elem::Link` type has changed from a struct to an
enum, with variants for normal links (containing the struct's previous
information) and autolinks.

This feels a bit cumbersome, but I think it correctly models the
difference between normal links and autolinks. The alternatives I
considered were:

- Not encoding autolink-ness at all, and instead just inferring it from
  the link's display text, title, and URL. This works in many cases, but
  would also convert intentionally-not-autolinks into autolinks. It also
  feels like autolinks shouldn't participate in link transformation.
- Adding a variant to the `LinkReference` enum for `Autolink` (or
  alternatively, adding an is-autolink flag to `LinkReference::Inline`).
  This would be a smaller change, but still an API-breaking one. Worse,
  the autolink implies certain invariants (that the title is `None`, and
  that the display text is just the plaintext of the URL) that would be
  impossible to enforce without making the types' fields non-public, in
  contrast to the other types in the module which are all PODs. I would
  therefore have to define what happens if an autolink has a title or
  incorrect display text, and that would inevitably lead towards "autolink
  is just a suggestion", which feels like a bug waiting to happen.

All in all, it felt like it was better to just be explicit about
autolinks, and how they're different from normal links.

This change resolves #297.

## Breaking change

The `mdq::md_elem::elem::Link` type has changed from a struct to an
enum, with variants for normal links (containing the struct's previous
information) and autolinks.