pass node slices around, instead of iterators
In service of #390. For that change, we'll want the set of all
reference-style links (e.g. `[1]`) before we start writing markdown.
This lets us do that without having to collect the iterator, which is
especially nice since the iterator is itself just coming from a vec
that's sliceable.

## Breaking change

- `mdq::output::MdWriter::write` takes an `&[MdElem]` instead of an
`impl IntoIterator<Item = &MdElem>`