simplify full-link transforms
If the original link was something like [123][a], respect that, and
don't renumber itt. Similarly, if it was [123][4], we assume that the
user is okay with that slight weirdness, and keep it (but renumber the
"4").

This does lead to a situation where ``[123][123]`` could bre renumbered
to a weirder [123][4]. I'm okay with that for now.