simplify link transformation
Remove `LinkTransform::Reference`, replacing it with
`LinkTransform::NeverInline`. This keeps collapsed and shortcut links as
they were (unlike `Reference`, which replaced them with `[full
references][1]`).

Additionally, collapsed and shortcut links whose display text is just a
number (e.g. `[1][]` or `[2]`) will now reserve that number, rather than
having it be renumbered as a full-style link. Otherwise, the link
transformer would change something like `[3]` to `[3][1]` (if that
shortcut `[3]` link happened to be the first link in the document),
which would result in confusing markdown. Full-style links like `[some
text][4]` will still be renumbered.

This simplifies link transformation, and also provides for a nicer user
experience. Resolves #390.

## Breaking changes

- `LinkTransform::Reference` has been removed